move tracefs
