+++ utils_test.go
@@ -3,7 +3,6 @@ package manager
 import (
 	"bytes"
 	"fmt"
-	"runtime"
 	"testing"
 )
 
@@ -106,22 +105,5 @@ func TestGetSyscallFnNameWithKallsyms(t *testing.T) {
 				t.Errorf("expected %s, got %s", testEntry.expected, res)
 			}
 		})
-	}
-}
-
-func BenchmarkFindFilterFunction(b *testing.B) {
-	var needle string
-	switch runtime.GOARCH {
-	case "arm64":
-		needle = "__arm64_sys_open"
-	default:
-		needle = "__x64_sys_open"
-	}
-
-	for i := 0; i < b.N; i++ {
-		_, err := FindFilterFunction(needle)
-		if err != nil {
-			b.Error(err)
-		}
 	}
 }
